requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.12
faiss-cpu>=1.8.0
onnx>=1.16.0
onnxruntime>=1.20.0
numpy>=1.26
pillow>=10.0
anomalib>=1.1.0
